requirements.txt
aniso8601==7.0.0
Babel==2.7.0
blinker==1.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==2.2.7
django-appconf==1.0.3
django-filter==2.2.0
django-imagekit==4.0.2
djangorestframework==3.10.3
Flask==1.1.1
Flask-Admin==1.5.4
Flask-BabelEx==0.9.3
Flask-GraphQL==2.0.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
graphene==2.1.8
graphene-django==2.6.0
graphene-django-extras==0.4.8
graphene-sqlalchemy==2.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
passlib==1.7.1
pilkit==2.0
Pillow==6.2.1
pipenv==2018.11.26
promise==2.2.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
speaklater==1.3
SQLAlchemy==1.3.10
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.7
virtualenv-clone==0.5.3
Werkzeug==0.16.0
WTForms==2.2.1
